requirements.txt
streamlit>=1.31.0
colorama>=0.4.6
openai
azure.storage.blob
azure.search.documents==11.5.2
python-dotenv
azure-core
azure.identity